# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export type EmmausConversationRequestContextEntryPoint = typeof EmmausConversationRequestContextEntryPoint[keyof typeof EmmausConversationRequestContextEntryPoint];


export const EmmausConversationRequestContextEntryPoint = {
  bible: 'bible',
  walk: 'walk',
  journeys: 'journeys',
  sermons: 'sermons',
  personal: 'personal',
  standalone: 'standalone',
} as const;

/**
 * Optional context about where the user came from
 */
export type EmmausConversationRequestContext = {
  entryPoint?: EmmausConversationRequestContextEntryPoint;
  userId?: string;
  conversationId?: string;
  userName?: string;
  bookId?: string;
  bookName?: string;
  chapter?: number;
  chapterHeading?: string;
  verseText?: string;
  journeyId?: string;
  journeyTitle?: string;
  currentDay?: number;
  sermonId?: string;
  sermonTitle?: string;
  scriptureReference?: string;
};

export type EmmausConversationRequestHistoryItemRole = typeof EmmausConversationRequestHistoryItemRole[keyof typeof EmmausConversationRequestHistoryItemRole];


export const EmmausConversationRequestHistoryItemRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export type EmmausConversationRequestHistoryItem = {
  role: EmmausConversationRequestHistoryItemRole;
  content: string;
};

export interface EmmausConversationRequest {
  message: string;
  /** Optional context about where the user came from */
  context?: EmmausConversationRequestContext;
  history?: EmmausConversationRequestHistoryItem[];
}

export interface EmmausConversationStub {
  id: string;
  userId: string;
  title: string;
  entryPoint: string;
  createdAt: string;
  updatedAt: string;
  messageCount: number;
}

export type EmmausMessageRole = typeof EmmausMessageRole[keyof typeof EmmausMessageRole];


export const EmmausMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface EmmausMessage {
  id: string;
  conversationId: string;
  userId: string;
  role: EmmausMessageRole;
  content: string;
  promptVersion: string;
  entryPoint: string;
  safetyChecked: boolean;
  createdAt: string;
}

export interface EmmausMemoryInput {
  userId: string;
  content: string;
  source?: string;
}

export interface EmmausMemory {
  id: string;
  userId: string;
  content: string;
  source: string;
  approved: boolean;
  createdAt: string;
}

export interface EmmausError {
  error: string;
}

export interface HealthStatus {
  status: string;
}

export type ListEmmausConversationsParams = {
userId?: string;
};

export type DeleteEmmausMemoryParams = {
userId?: string;
};